refactor: reorganize packages and adjust imports across Matching and UserSetting modules
