+++ src/main/java/com/yeoljeong/tripmate/matching/domain/entity/Matching.java
@@ -1,10 +1,6 @@
 package com.yeoljeong.tripmate.matching.domain.entity;
 
-import com.yeoljeong.tripmate.domain.BaseAuditEntity;
 import com.yeoljeong.tripmate.matching.domain.constants.MatchingStatus;
-import com.yeoljeong.tripmate.matching.domain.entity.vo.Location;
-import com.yeoljeong.tripmate.matching.domain.entity.vo.MatchingPeriod;
-import com.yeoljeong.tripmate.matching.domain.entity.vo.MatchingSetting;
 import jakarta.persistence.Column;
 import jakarta.persistence.Embedded;
 import jakarta.persistence.Entity;
@@ -16,7 +12,6 @@ import jakarta.persistence.Id;
 import jakarta.persistence.Table;
 import java.util.UUID;
 import lombok.AccessLevel;
-import lombok.Getter;
 import lombok.NoArgsConstructor;
 import org.hibernate.annotations.UuidGenerator;
