+++ src/main/java/com/yeoljeong/tripmate/matching/infrastructure/persistence/jpa/SpringDataMatchingRepository.java
@@ -1,4 +1,4 @@
-package com.yeoljeong.tripmate.matching.infrastructure.repository.jpa;
+package com.yeoljeong.tripmate.matching.infrastructure.persistence.jpa;
 
 import com.yeoljeong.tripmate.matching.domain.entity.Matching;
 import java.util.UUID;
