+++ src/main/java/com/yeoljeong/tripmate/usersetting/application/dto/result/UserSettingResult.java
@@ -1,4 +1,4 @@
-package com.yeoljeong.tripmate.usersetting.application.result;
+package com.yeoljeong.tripmate.usersetting.application.dto.result;
 
 import com.yeoljeong.tripmate.usersetting.domain.entity.UserSetting;
 import com.yeoljeong.tripmate.usersetting.domain.entity.constants.Gender;
