+++ src/main/java/com/yeoljeong/tripmate/usersetting/presentation/dto/response/MbtiResponse.java
@@ -1,6 +1,6 @@
-package com.yeoljeong.tripmate.usersetting.presentation.response;
+package com.yeoljeong.tripmate.usersetting.presentation.dto.response;
 
-import com.yeoljeong.tripmate.usersetting.application.result.MbtiResult;
+import com.yeoljeong.tripmate.usersetting.application.dto.result.MbtiResult;
 
 public record MbtiResponse(
 	String ie,
